requirements.txt
py~=1.10.0
allure_pytest~=2.9.45
colorama~=0.4.4
pip~=19.2.3
configparser~=5.0.2
attrs~=20.3.0
toml~=0.10.2
click~=7.1.2
requests~=2.25.1
appdirs~=1.4.4
packaging~=20.8
six~=1.16.0
pipenv~=2021.5.29
distlib~=0.3.3
setuptools~=41.2.0
MarkupSafe~=1.1.1
urllib3~=1.26.2
certifi~=2021.10.8
chardet~=4.0.0
pytest~=6.2.5
idna~=2.10
pyparsing~=2.4.7
crayons~=0.4.0
pluggy~=0.13.1
iniconfig~=1.1.1
atomicwrites~=1.4.0
numpy~=1.21.3
virtualenv~=20.8.1
filelock~=3.3.0
platformdirs~=2.4.0
selenium~=3.141.0
Pillow~=8.4.0
assertpy~=1.1